+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # changelog
 
+## [1.1.0] (2025-04-06)
+
+* added config entries for protection timer and glow color, with six different options available for glow timer
+* fixed incorrect bepinex version
+* tidied up some stuff
+
 ## [1.0.6] (2025-03-30)
 
 * further optimized and reduced unnecessary color capture due to switching to emission-only